feat: move controller models to subpackage for consistency
