+++ bot/src/main/java/edu/java/bot/model/controller/LinkUpdateRequest.java
@@ -1,4 +1,4 @@
-package edu.java.bot.model;
+package edu.java.bot.model.controller;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
