+++ bot/src/main/java/edu/java/bot/model/controller/exceptions/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package edu.java.model.exception;
+package edu.java.bot.model.controller.exceptions;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
