+++ scrapper/src/main/java/edu/java/model/controller/RemoveLinkRequest.java
@@ -1,10 +1,10 @@
-package edu.java.model;
+package edu.java.model.controller;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.net.URI;
 
-public record AddLinkRequest(
+public record RemoveLinkRequest(
     @Schema(name = "link", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    URI link
+    URI request
 ) {
 }
